feat: Enhance mobile card column visibility in SuperTable
Introduces `maxVisibleColumns` prop to `SuperTable` and `TableRowMobileCard` to control the number of columns initially displayed in the mobile card view.

Renames `cardPriority` to `cardOverflowPriority` in `ColumnDef` to better reflect its purpose: defining the priority for columns that are moved into the "Show More" section when `maxVisibleColumns` is applied. Columns without `cardOverflowPriority` are now prioritized for initial visibility.